+++ R/arima.R
@@ -10,41 +10,34 @@ make_arima <- function(phi, theta, delta) {
     ret
 }
 
-#' Simulate from arima models
+#' Apply ARIMA recursions
 #'
-#' The innovations `innov` are assumed to include the burn-in phase. Only the
-#' last `nsim` simulations are returned.
-#' The implementation is based on `stats::arima.sim`.
-#' @param object a fitted arima model
-#' @param nsim positive integer. Length of time series to be generated.
-#' @param innov vector of innovations.
+#' This implementation is based on `stats::arima.sim`.
+#' @param object an `arima` object
+#' @inheritParams inn2ts
 #' @param ... not used.
 #' @return a [tsibble][tsibble::tsibble-package] of class "arima_ts". Column
-#'      "value" corresponds to the simulated time series and column "date" has
+#'      "ts" corresponds to the simulated time series and column "date" has
 #'      dummy dates for each observation.
 #' @export
 #' @examples
 #' # simulate from an AR(1) model
 #' mdl <- make_arima(phi = 0.3, theta = 0, delta = 0)
-#' simults(mdl, nsim = 100, innov = rnorm(200))
-simults.arima <- function(object, nsim, innov, ...) {
+#' inn2ts(mdl, inn = rnorm(200))
+inn2ts.arima <- function(object, inn, ...) {
+    len_inn <- length(inn)
     phi <- object$phi
     theta <- object$theta
     delta <- object$delta
     len_delta <- length(delta)
     len_phi <- length(phi)
-    if (!is.numeric(innov)) rlang::abort("`innov` is not numeric")
-    if (vctrs::vec_size(innov) <= nsim) {
-        rlang::abort(
-            "Length of `innov` should be more than `nsim`."
-        )
-    }
-    ret <- innov
+    if (!is.numeric(inn)) rlang::abort("`inn` is not numeric")
+    ret <- inn
     # apply theta
     if (length(theta)) {
         ret <- stats::filter(ret, c(1, theta), sides = 1L) |>
-            stats::na.omit() |>
             as.numeric()
+        ret[1:length(theta)] <- 0
     }
     # apply phi
     if (len_phi) {
@@ -57,8 +50,8 @@ simults.arima <- function(object, nsim, innov, ...) {
             as.numeric()
     }
     tsibble::tsibble(
-        date = Sys.Date() + 1:nsim,
-        value = utils::tail(ret, nsim),
+        date = Sys.Date() + 1:len_inn,
+        ts = ret,
         index = date
     ) |>
         tsibble::new_tsibble("arima_ts")
@@ -67,14 +60,12 @@ simults.arima <- function(object, nsim, innov, ...) {
 
 #' Fitted residuals for ARIMA
 #'
-#' @inheritParams fitted_resid
-#' @param new_data a [tsibble][tsibble::tsibble-package] of univariate
-#'      time series on which residuals are to be calculated.
-#' @param ... not used
-#' @return a [tsibble][tsibble::tsibble-package] of fitted residuals of class
-#'      `fitted_resid_ts`.
+#' @inheritParams ts2inn
+#' @param ... not used.
+#' @return a [tsibble][tsibble::tsibble-package] where column `inn` contains the
+#'      calculated innovations.
 #' @export
-fitted_resid.arima <- function(object, new_data, ...) {
+ts2inn.arima <- function(object, ts, ...) {
     if (!inherits(object, "arima")) rlang::abort("Not arima object.")
     phi <- object$phi
     theta <- object$theta
@@ -82,8 +73,8 @@ fitted_resid.arima <- function(object, new_data, ...) {
     len_delta <- length(delta)
     len_phi <- length(phi)
     # Extract response column
-    new_data_response_var <- tsibble::measured_vars(new_data)
-    ret <- new_data[[new_data_response_var]]
+    ts_response_var <- tsibble::measured_vars(ts)
+    ret <- ts[[ts_response_var]]
     ret <- c(rep(0, len_delta + len_phi), ret)
     # Differencing
     if (len_delta) {
@@ -99,8 +90,7 @@ fitted_resid.arima <- function(object, new_data, ...) {
     if (length(theta)) {
         ret <- stats::filter(ret, -theta, method = "recursive")
     }
-    new_data |>
-        dplyr::mutate(".resid" = ret) |>
-        dplyr::select(-!!new_data_response_var) |>
-        tsibble::new_tsibble(class = "fitted_resid_ts")
+    ts |>
+        dplyr::mutate("inn" = ret) |>
+        dplyr::select(-!!ts_response_var)
 }